fixed all register controller

+++ src/utils/cloudinary.js
@@ -1,21 +1,33 @@
-import {v2 as cloudinary}from "cloudinary"
-import fs from "fs"
-cloudinary.config({ 
-    cloud_name:process.env.CLOUDINARY_CLOUD_NAME, 
-    api_key: process.env.CLOUDINARY_CLOUD_NAME, 
-    api_secret: process.env.CLOUDINARY_API_SECRET 
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
+import dotenv from "dotenv"
+dotenv.config()
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
 });
-const uploadOnCloudinary=async(localFilePath)=>{
-    try{
-        if (!localFilePath)return null
-        const response=await cloudinary.upload(localFilePath,{resource_type:'auto'})
-        console.log('file uploaded successfully on cloudinary')
-        return response
 
-    }catch(error){
-        fs.unlinkSync(localFilePath)
-        return null
+const uploadOnCloudinary = async (localFilePath) => {
+  try {
+    if (!localFilePath) return null;
+
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto",
+    });
+
+    // console.log("File uploaded successfully on cloudinary");
+    fs.unlinkSync(localFilePath)
+    return response;
+  } catch (error) {
+    console.error("Cloudinary upload error:", error);
+
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
     }
-}
-cloudinary.v2.uploader.upload()
-export {uploadOnCloudinary}
+
+    return null;
+  }
+};
+
+export { uploadOnCloudinary };
